Attackers now move to gain LoS.
When attacking a target, attacker will now move to gain line of sight to the target and start shooting.

TODO: There's a bug in the LoS check -- RaycastHits aren't guarenteed to be in any particular order. so we need to find the closest hit and check if that is our target, rather than just checking the first hit.